streamlit/requirements.txt
aioice==0.7.6
aiortc==1.3.2
altair==4.2.0
attrs==22.1.0
av==9.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.10.0
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
glob2==0.7
google-crc32c==1.5.0
idna==2.10
importlib-metadata==5.0.0
importlib-resources==5.10.0
Jinja2==3.1.2
jsonschema==4.17.0
kiwisolver==1.3.1
MarkupSafe==2.1.1
matplotlib==3.6.2
netifaces==0.11.0
numpy==1.23.4
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==10.0.0
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.13.0
pylibsrtp==0.7.1
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
roboflow==0.2.20
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.0
streamlit-webrtc==0.44.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.6
validators==0.20.0
watchdog==2.1.9
wget==3.2
zipp==3.10.0
pyserial
